requirements.txt
# Core data
pandas>=2.0.0
geopandas>=0.14.0
openpyxl>=3.1.0
pyarrow>=14.0.0

# Spatial
shapely>=2.0.0
pyproj>=3.6.0
geopy>=2.4.0

# LLM APIs
openai>=1.30.0
anthropic>=0.28.0
python-dotenv>=1.0.0

# Web scraping
requests>=2.31.0
beautifulsoup4>=4.12.0

# Telegram
telethon>=1.36.0

# ML
lightgbm>=4.3.0
scikit-learn>=1.4.0
imbalanced-learn>=0.12.0

# Visualisation
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.20.0
folium>=0.16.0
streamlit>=1.35.0
streamlit-folium>=0.20.0

# Notebooks
notebook>=7.0.0
ipykernel>=6.29.0
